Changed floats to doubles, Fixed instantiation issue

+++ src/collision/collisionObject.h
@@ -11,7 +11,7 @@ using namespace nanogui;
 
 class CollisionObject {
 public:
-  virtual void collide(Particle &p, float cr, float delta_t) = 0;
+  virtual void collide(Particle &p, double cr, double delta_t) = 0;
 
 private:
   double friction;
